reuse previous occurances of functions
This fixes the issue when a function is referenced several times in the
data, and docca creates an instance for each reference, but the index
only contains the first instance. That in turn would lead to several
attributes of other instances not being set. # Please enter the commit
message for your changes. Lines starting # with '#' will be ignored, and
an empty message aborts the commit. # # On branch fix/duplicate-entities