ensure related functions are scoped correctly
The assumption is that related functions are "members" of classes they
relate to. This assumption could be violated with recent Doxygen if the
first compound to have them as members is their namespace. Now Docca
makes sure to remove such functions from every scope but one, and also
makes sure that last scope is a class.